Added DataLoader code
Added ipynb file that creates and describes the dataloaders for both the test and train set.

An important note - as they stand now, the lengths of the reviews are variable. For simplicity, it may be worth it to explore limiting them to a max length (I was thinking 256 words?) and then zero padding reviews that are shorter up to 256.